Add Modules to PRODUCT_PACKAGES list
Modules which do not have LOCAL_MODULE_TAGS defined default
to "user" tag. This results in errors and warning messages
in the Android build system in Gingerbread. These errors and
warning messages can be eliminated by adding "optional" tag.
When we add "optional" tag to any of the modules, the module
should be part of PRODUCT_PACKAGES list, to be available in
the final images that we want to load on to the target.
This change adds modules to PRODUCT_PACKAGES list.

Change-Id: I28230a916e116ee9621c6559f53bd9ff542a291f