init.qcom.post_boot.sh: Disable core_control when setting cpu params
Kernel thermal migitation restricts the number of cores that can be
onlined, when in certain thermal situations. To set cpufreq parameters,
the core should be online. If the core_control feature is not disabled,
the cores cannot be brought online and the following commands will fail.
Enable core_control after the parameters are set.

CRs-Fixed: 460083
Change-Id: Icb8f073547483a1acdf97937cc1b6fae3add33a6