init: post_boot: create a persistent files for mpdecision
Mpdecision may need to store data which should be persistent
across restarts and it uses this data to reconfigure the init scripts.
When mpdecision brings secondary cores online, it stores the
low power mode setting in local variables and disables them in init
script. If mpdecision gets restarted at this moment there is no way
it can reconfigure low power mode settings in init scripts. In this
context it stores the data in persistent storage and restore it.

CRs-fixed: 492804

Change-Id: Ie5de2c8a79a008d87be3ca66888500ab406c8f3d