Fix subsystem initialization issue.
This change fixes the subsystem initializatio
issue. So we need to create soft links
based on the hw revision. This is being done
as same non-hlos.bin has multiple modems and
we are doing fixups here. If non-hlos.bin has
single modem (with right names) then this is
not needed.

Change-Id: I5cff0703b5bc88715b926c21afd6d2d634882a2e