CameraApp: Flash mode settings array values reordered
Camera application decides to enable or disable the settings checkbox
based on the index of the array element. For array element at index 0
checkbox will be uncheked and for element at index 1 checkbox will be
checked. So changed the order of the elements such that Off is at
index 0 and On or torch is at index 1.

Change-Id: I8ff22708dee357b00a3f994f06188d363946c980
CRs-Fixed: 566779