CPUFreq: Set optimal frequency for msm8226.
The ondemand governor has the ability to adapt to multicore
targets to render better scaling. This change sets one of such
parameters i.e optimal frequency to the nominal max frequency
to achieve best perf/power. Along with that, this change also
tunes the down_differential_multi_core to a reasonable value.

CRs-fixed: 561456
Change-Id: If0ea8e2f663982201f7d680144def44fbfd62c50