CPUFreq: Set optimal frequency for msm8610
The ondemand governor has the ability to adapt to multicore
targets to render better scaling. This change sets one of such
parameters i.e optimal frequency to the nominal max frequency
to achieve best perf/power. Along with that, this change also
tunes the down_differential_multi_core to a reasonable value.

Change-Id: I4fe76224f0166a3a6db2a79c18637e0ab760ceac
CRs-fixed: 561456