device: usb: add changes for diag to work with rndis composition
we use persist.sys.usb.config.extra to specify if we want to use diag with
rndis interface earlier userspace used to read this property to check
whether diag is to be included in composition and used to set appropriate
composition.

Now we plan to make this check and change in device specific files. This
change checks whenever there is a composition with rndis and
persist.sys.usb.config.extra indicates to include diag we set the
usb.sys.config to reflect this indication. User now need to set extra
property to diag and based on the baseband information we can figure out
clients representing diag interface.

Change-Id: I3257f47ed42b3ad4fea68a6deb45c27944614558