init.qcom.rc: Increase tcp buffer sizes
Applications which deal with high data rates were observed to
not be able to consume data fast enough. This was causing queueing
in the TCP stack. Since the queues were not large enough, it caused
dropping of packets and subsequent drops of packets in out of order
queues. As a result, TCP downlink and bi-directional throughput was
severely impacted. The TCP buffer sizes are increased to achieve
peak throughputs in these scenarios.

CRs-fixed: 630093
Change-Id: Ifc5a10bb683e0c1f044f1df5d8466b4ecfa846ba