USB: init.qcom.usb.sh: Add platform name msm8916_32 in default composition
If ADB or NONE composition is set, USB init script
switch the composition to default composition which
is different on MSM8916 with PID 9091.

Target name has recently changed from msm8916 to
msm8916_32. Because of this, USB init script is not
able to move to 9091 composition.  Add this target
name check to fix this issue.

CRs-Fixed: 646400
Change-Id: I7107ad2ef9d0f301813074da439f62938886d881