init.qcom.usb.sh: Initialize persist.sys.usb.config.extra property
commit c6a666ce "init.qcom.usb.sh: Remove the diag_extra section"
removed the section that sanitizes the persist.sys.usb.config.extra
property. However, if it is unset, the init.qcom.usb.rc rules will
fail to parse the empty property upon USB composition switch to
RNDIS. Instead, it needs to at least be initialized to 'none' if
not already set.

Change-Id: I98c08180bcbe791667d0ccb866a11796a7debc12