USB: init.qcom.usb.sh: Add 64-bit kernel target name
Default composition on MSM8916 is different from
other targets and this distinction to switch back
to default composition is made on the basis of
target name.

Because of multiple variants of 32-bit and 64-bit
kernel and userspace, the target names are
assigned different for these builds. Parse the first
seven letters in target name. These seven letters
cover for 'msm' and XXXX digits. A 64-bit build
target name msm8916_32_64k is parsed to msm8916
and msm8916_32 is parsed to msm8916.

CRs-Fixed: 674082
Change-Id: If17afce5612c2a16515a3d2b30eab381e1e92076