msm8939: Create service for msm_irqbalance and start it
Create the appropriate service for msm_irqbalance daemon and
start the service if we are running on msm8939.

The msm_irqbalance daemon starts up as root and then lower its
privileges immediately to user nobody while holding on to
capability to read/write files on the file system (proc).

The msm_irqbalance daemon balances IRQs among available cores to ensure
even IRQ load on each core. The daemon also moves IRQs away from cores
that are configured to not handle any IRQs.

Change-Id: I728c91b31bb3cfbdd14c630d3a1d93a0fed18460