Bluetooth: Change group ID to net_bt_stack for bluetooth files
Change the group ID from bluetooth to net_bt_stack to enable BT in
guest in login. Bluetooth UID is OR'ed with App id and it keeps
incrementing with every new logg in as guest. As bluetooth uid
is prefixed with app Id, bluetooth owned files with group Id as
bluetooth cannot be accessed.

Change-Id: I0a83da59b8c6a5d90c63f962f641cc780dc8e358
CRs-fixed: 770726