post_boot:msm8916:8939:8909 sched policy to consider avg of windows
change scheduler policy on L update to consider average of
last 5 windows instead of max(last window, average of 5 windows).
It was same on KK (i.e. average of 5 windows) and changed on L
due to sched code updates. so fixing this on L build as well.

Also,remvoing redundant setting of writing to sched_ravg_hist_size.

Conflicts:
	rootdir/etc/init.qcom.post_boot.sh

Change-Id: Idc882fc40b9b963fa370c7ce920040f5e7c475c9