common: Add the packkernelimg tool
This tool adds a header to the start of the kernel image which
indicates the size of the uncompressed kernel. It also appends
the device tree blobs to the end of the image. The header at the
start is used by the bootloader to calculate the start of the device
tree blob.

Change-Id: Ifb5428f42ee3344b37bdb2fe34c7e86afcd6a63f