Fix permissions to internal ramdump and dload_mode
System apps such as RIDL/LogKit need to be able to enable internal
mini crash dumps and to collect the dump. rw permissions are
needed for dload_mode node to enable/disable feature. rw permissions are
needed for ramdump to read the ramdump partition and to mark it as read.

This fix ensures the dev nodes are properly set for the rawdump partition.
The original fix did have this set correctly but somehow was changed.

Change-Id: Icbe811cbe4394f47400bbadd1bb8774d59bd4c28