Spyro: suppress cpu uevents in post_boot
Ignore CPU uevents as the sysfs nodes for CPU subsystem are persistent
across hotplugs and we can ignore processing the CPU sysfs nodes during
CPU online uevent notifications.

If we don't suppress the CPU uevents, the sysfs processing in ueventd
takes a considerable amount of CPU load on the newly online CPU
preventing the actual CPU load - which triggered the new CPU to be
brought online - from executing.

Change-Id: Iee4e8d4a82d36bb74fbab1342ed83d40d84105d3